refactor: move utils to services
